git-svn-id: svn+ssh://188.165.179.87/var/svn/mouf/trunk@435 20a9f3e9-cf9c-4ed3-b8ae-ea11efa85a44

+++ MoufPackageManager.php
@@ -604,7 +604,7 @@ class MoufPackageManager {
 	 * @return string The path to the ZIP file.
 	 */
 	public function getZipFilePath(MoufPackage $moufPackage) {
-		$packageDir = ROOT_PATH.$moufPackage->getPackageDirectory();
+		$packageDir = ROOT_PATH."plugins/".$moufPackage->getPackageDirectory();
 		$zipFileName = $moufPackage->getDescriptor()->getName()."-".$moufPackage->getDescriptor()->getVersion().".zip";
 		return $packageDir.'/../'.$zipFileName;
 	}
@@ -619,15 +619,17 @@ class MoufPackageManager {
 	 */
 	public function compressPackage(MoufPackage $moufPackage) {
 		
-		$packageDir = ROOT_PATH.$moufPackage->getPackageDirectory();
+		$packageDir = ROOT_PATH."plugins/".$moufPackage->getPackageDirectory();
 		
 		//echo $packageDir;
 						
 		//$zipFileName = $moufPackage->getDescriptor()->getName()."-".$moufPackage->getDescriptor()->getVersion().".zip";
 		$zipFilePath = $this->getZipFilePath($moufPackage);
 		
-		
 		if (file_exists($zipFilePath)) {
+			if (!is_writable($zipFilePath)) {
+				throw new MoufException("Cannot delete ZIP file ".$zipFilePath);
+			}
 			unlink($zipFilePath);
 		}
 
@@ -636,6 +638,11 @@ class MoufPackageManager {
 		
 		// create object
 		$zip = new ZipArchive();
+		
+		$dirname = dirname($zipFilePath);
+		if (!is_writable($dirname)) {
+			throw new MoufException("Cannot create ZIP file in directory ".$dirname.". Please check the directory rights.");
+		}
 		
 		// open output file for writing
 		if ($zip->open($zipFilePath, ZIPARCHIVE::CREATE) !== TRUE) {
